internal: move CPU detection to separate package
